#214 [feat] Retrofit Builter를 싱글톤으로 변경
