+++ app/src/main/java/com/gojol/notto/network/GithubService.kt
@@ -1,10 +1,6 @@
 package com.gojol.notto.network
 
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Response
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
 import retrofit2.http.GET
 import retrofit2.http.Path
 
@@ -15,23 +11,4 @@ interface GithubService {
         @Path("owner") owner: String,
         @Path("repo") repo: String
     ): Response<GithubResponse>
-
-    companion object {
-        fun getService(baseUrl: String): GithubService {
-
-            val logger = HttpLoggingInterceptor()
-            logger.level = HttpLoggingInterceptor.Level.BASIC
-
-            val client = OkHttpClient.Builder()
-                .addInterceptor(logger)
-                .build()
-
-            return Retrofit.Builder()
-                .baseUrl(baseUrl)
-                .client(client)
-                .addConverterFactory(GsonConverterFactory.create())
-                .build()
-                .create(GithubService::class.java)
-        }
-    }
 }
